requirements.txt
streamlit
google-generativeai
python-dotenv
PyMuPDF
requests
duckduckgo-search
huggingface
transformers
langchain-core
faiss-cpu
watchdog
langchain_community
sentence-transformers
